+++ siddhi-testsuite/src/main/resources/archetype-resources/component/src/test/java/integrationtests/IntegrationTestsOf__classNameOfTestsuite__.java
@@ -1,23 +1,18 @@
 package ${package}.integrationtests;
 
 import com.google.common.io.Resources;
-import io.siddhi.core.SiddhiAppRuntime;
-import io.siddhi.core.SiddhiManager;
 import io.siddhi.core.exception.ConnectionUnavailableException;
-import io.siddhi.core.stream.input.InputHandler;
 import io.siddhi.distribution.test.framework.MySQLContainer;
 import io.siddhi.distribution.test.framework.NatsContainer;
 import io.siddhi.distribution.test.framework.SiddhiRunnerContainer;
 import io.siddhi.distribution.test.framework.util.DatabaseClient;
 import io.siddhi.distribution.test.framework.util.NatsClient;
-import io.siddhi.extension.io.nats.sink.NATSSink;
-import io.siddhi.extension.map.json.sinkmapper.JsonSinkMapper;
+import ${package}.AbstractTests;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.testcontainers.containers.Network;
 import org.testcontainers.containers.output.OutputFrame;
 import org.testcontainers.containers.output.Slf4jLogConsumer;
-import org.testcontainers.containers.output.WaitingConsumer;
 import org.testng.Assert;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
@@ -29,31 +24,17 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 
 /**
  * Class for integration testing.
  *
  */
-public class IntegrationTestsOf${classNameOfTestsuite} {
+public class IntegrationTestsOf${classNameOfTestsuite} extends AbstractTests {
 
     private static final Logger logger = LoggerFactory.getLogger(IntegrationTestsOf${classNameOfTestsuite}.class);
-
-    MySQLContainer mySQLContainer;
-    NatsContainer natsContainer;
-    SiddhiRunnerContainer siddhiRunnerContainer;
-
-    NatsClient natsClient;
-    WaitingConsumer siddhiLogConsumer = new WaitingConsumer();
-
-    private String natsClusterId = "TemperatureCluster";
-    private String natsUrl = "nats-streaming";
-    private String natsInputDestination = "Temp-Alert-App_DeviceTempStream";
-    private String natsOutputDestination = "Temp-Alert-App_AlertStream";
 
     private static final String DATABSE_NAME = "TemperaureDB";
     private static final String DATABSE_HOST = "mysqldb";
@@ -105,14 +86,6 @@ public class IntegrationTestsOf${classNameOfTestsuite} {
                 NATS_OUTPUT_DESTINATION);
     }
 
-    public void setClusterConfigs(String natsClusterId, String natsUrl, String natsInputDestination,
-                                  String natsOutputDestination) {
-        this.natsClusterId = natsClusterId;
-        this.natsUrl = natsUrl;
-        this.natsInputDestination = natsInputDestination;
-        this.natsOutputDestination = natsOutputDestination;
-    }
-
     @AfterClass
     public void shutdownCluster() {
         if (natsContainer != null) {
@@ -150,99 +123,5 @@ public class IntegrationTestsOf${classNameOfTestsuite} {
                 resultSet.close();
             }
         }
-    }
-
-    @Test
-    public void testMessageConsumption() throws InterruptedException, IOException, TimeoutException {
-        natsClient.publish(natsInputDestination, "{\n" +
-                "    \"event\": {\n" +
-                "        \"type\": \"dummyType\",\n" +
-                "        \"deviceID\": \"dummyDeviceID\",\n" +
-                "        \"temp\": 50.0,\n" +
-                "        \"roomID\": \"dummyRoomID\"\n" +
-                "    }\n" +
-                "}");
-        try {
-            siddhiLogConsumer.waitUntil(frame ->
-                            frame.getUtf8String().contains("data=[dummyType, dummyDeviceID, 50.0, dummyRoomID]"),
-                    50, TimeUnit.SECONDS);
-        } catch (TimeoutException e) {
-            Assert.fail("Message consumption acknowledgement is not available in Siddhi Runner logs.");
-        }
-    }
-
-    @Test
-    public void testMessagePublishingWithSiddhi() throws InterruptedException {
-
-        SiddhiManager siddhiManager = new SiddhiManager();
-        siddhiManager.setExtension("nats-sink", NATSSink.class);
-        siddhiManager.setExtension("sink-mapper-json", JsonSinkMapper.class);
-        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(
-                "@App:name(\"PublishToNatsApp\")\n" +
-                "\n" +
-                "@App:description('Publishes events to nats topic')\n" +
-                "\n" +
-                "@sink(type='nats', cluster.id='" + natsClusterId + "', " +
-                "destination = '" + natsInputDestination + "', " +
-                "bootstrap.servers = '" + natsUrl + "' ,@map(type='json'))\n" +
-                "define stream DeviceTemperatureStream (type string, deviceID string, " +
-                "temp double, roomID string);");
-        siddhiAppRuntime.start();
-        InputHandler inputHandler = siddhiAppRuntime.getInputHandler("DeviceTemperatureStream");
-        inputHandler.send(new Object[]{"fooType", "001", 60, "202"});
-        try {
-            siddhiLogConsumer.waitUntil(frame ->
-                            frame.getUtf8String().contains("data=[fooType, 001, 60.0, 202]"),
-                    50, TimeUnit.SECONDS);
-        } catch (TimeoutException e) {
-            Assert.fail("Message consumption acknowledgement is not available in Siddhi Runner logs.");
-        } finally {
-            siddhiManager.shutdown();
-        }
-    }
-
-    @Test
-    public void testAppOutput() throws InterruptedException, IOException, TimeoutException {
-
-        NatsClient.ResultHolder resultHolder = new NatsClient.ResultHolder(1, 3);
-        NatsClient natsClient = new NatsClient(natsClusterId, "stan_test1",
-                natsUrl, resultHolder);
-        natsClient.connect();
-        natsClient.subscribeFromNow(natsOutputDestination);
-        natsClient.publish(natsInputDestination, "{\n" +
-                "    \"event\": {\n" +
-                "        \"type\": \"monitored\",\n" +
-                "        \"deviceID\": \"C001\",\n" +
-                "        \"temp\": 40.2,\n" +
-                "        \"roomID\": \"F2-Conference\"\n" +
-                "    }\n" +
-                "}");
-
-        natsClient.publish(natsInputDestination, "{\n" +
-                "    \"event\": {\n" +
-                "        \"type\": \"monitored\",\n" +
-                "        \"deviceID\": \"C001\",\n" +
-                "        \"temp\": 60.0,\n" +
-                "        \"roomID\": \"F2-Conference\"\n" +
-                "    }\n" +
-                "}");
-        natsClient.publish(natsInputDestination, "{\n" +
-                "    \"event\": {\n" +
-                "        \"type\": \"monitored\",\n" +
-                "        \"deviceID\": \"C001\",\n" +
-                "        \"temp\": 80.0,\n" +
-                "        \"roomID\": \"F2-Conference\"\n" +
-                "    }\n" +
-                "}");
-        natsClient.publish(natsInputDestination, "{\n" +
-                "    \"event\": {\n" +
-                "        \"type\": \"monitored\",\n" +
-                "        \"deviceID\": \"C001\",\n" +
-                "        \"temp\": 30.0,\n" +
-                "        \"roomID\": \"F2-Conference\"\n" +
-                "    }\n" +
-                "}");
-        Thread.sleep(10000);
-        Assert.assertTrue(((ArrayList<String>) resultHolder.waitAndGetResults()).get(0).contains("\"peakTemp\":80.0"));
     }
 }
